requirements.txt
langchain==0.0.1
langchain_community==0.0.1
langchain_google_genai==0.0.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
langchain-openai
pdfminer.six